+++ libandroid-navigation/src/test/java/com/mapbox/services/android/navigation/v5/snap/SnapToRouteTest.java
@@ -15,11 +15,13 @@ import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
 
 import static junit.framework.Assert.assertEquals;
+import static org.mockito.internal.matchers.text.ValuePrinter.print;
 
 @RunWith(RobolectricTestRunner.class)
 public class SnapToRouteTest extends BaseTest {
 
   private static final String MULTI_LEG_ROUTE_FIXTURE = "directions_two_leg_route.json";
+  private static final String SINGLE_STEP_LEG = "directions_three_leg_single_step_route.json";
 
     @Test
     public void getSnappedLocation_returnsProviderNameCorrectly() throws Exception {
@@ -162,6 +164,59 @@ public class SnapToRouteTest extends BaseTest {
   }
 
   @Test
+  public void getSnappedLocation_bearingWithSingleStepLegBeforeNextLeg() throws Exception {
+    DirectionsRoute routeProgress = buildMultipleLegRoute(SINGLE_STEP_LEG);
+    Snap snap = new SnapToRoute();
+    Location location = new Location("test");
+    location.setLatitude(38.8943771);
+    location.setLongitude(-77.0782341);
+    location.setBearing(20);
+
+    Location previousSnappedLocation = snap.getSnappedLocation(location, buildTestRouteProgress(
+            routeProgress,
+            0.8,
+            0.8,
+            200,
+            20,
+            0
+    ));
+
+    Location snappedLocation = snap.getSnappedLocation(location, buildTestRouteProgress(
+            routeProgress,
+            0.8,
+            0.8,
+            200,
+            21,
+            0
+    ));
+
+    // Latest snapped bearing should be used, because next lef is not containing enough steps
+    assertEquals(previousSnappedLocation.getBearing(), snappedLocation.getBearing());
+  }
+
+  @Test
+  public void getSnappedLocation_bearingNoBearingBeforeWithSingleStepLegBeforeNextLeg() throws Exception {
+    DirectionsRoute routeProgress = buildMultipleLegRoute(SINGLE_STEP_LEG);
+    Snap snap = new SnapToRoute();
+    Location location = new Location("test");
+    location.setLatitude(38.8943771);
+    location.setLongitude(-77.0782341);
+    location.setBearing(20);
+
+    Location snappedLocation = snap.getSnappedLocation(location, buildTestRouteProgress(
+            routeProgress,
+            0.8,
+            0.8,
+            200,
+            21,
+            0
+    ));
+
+    // Fallback to location bearing if no previous bearing was calculated.
+    assertEquals(location.getBearing(), snappedLocation.getBearing());
+  }
+
+  @Test
   public void getSnappedLocation_bearingEnd() throws Exception {
     DirectionsRoute routeProgress = buildMultipleLegRoute();
     Snap snap = new SnapToRoute();
@@ -169,6 +224,15 @@ public class SnapToRouteTest extends BaseTest {
     location.setLatitude(38.9623091);
     location.setLongitude(-77.0282631);
     location.setBearing(20);
+
+    Location lastSnappedLocation = snap.getSnappedLocation(location, buildTestRouteProgress(
+        routeProgress,
+        0.6,
+        0.6,
+        0.6,
+        14,
+        1
+    ));
 
     Location snappedLocation = snap.getSnappedLocation(location, buildTestRouteProgress(
         routeProgress,
@@ -179,11 +243,16 @@ public class SnapToRouteTest extends BaseTest {
         1
     ));
 
-    assertEquals(0f, snappedLocation.getBearing());
+    // Latest snapped bearing should be used, because no future steps are available
+    assertEquals(lastSnappedLocation.getBearing(), snappedLocation.getBearing());
   }
 
   private DirectionsRoute buildMultipleLegRoute() throws Exception {
-    String body = loadJsonFixture(MULTI_LEG_ROUTE_FIXTURE);
+    return buildMultipleLegRoute(MULTI_LEG_ROUTE_FIXTURE);
+  }
+
+  private DirectionsRoute buildMultipleLegRoute(String file) throws Exception {
+    String body = loadJsonFixture(file);
     Gson gson = new GsonBuilder().registerTypeAdapterFactory(DirectionsAdapterFactory.create()).create();
     DirectionsResponse response = gson.fromJson(body, DirectionsResponse.class);
     return response.routes().get(0);